fix: cleaning for notifications
